requirements.txt
flask
gunicorn
psycopg2-binary
groq
requests
python-dotenv
